% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nof1.wrap2.R
\name{find_raw_mean2}
\alias{find_raw_mean2}
\title{Summarizes the result from the model into json format}
\usage{
find_raw_mean2(Y, Treat, baseline, response)
}
\description{
Summarizes the result from the model into json format
}
