% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nof1.summary.R
\name{frequency_plot}
\alias{frequency_plot}
\title{Frequency plot for raw data}
\usage{
frequency_plot(nof1)
}
\arguments{
\item{nof1}{nof1 object created using nof1.data}
}
\description{
Frequency plot for raw data
}
