% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nof1.summary.R
\name{kernel_plot}
\alias{kernel_plot}
\title{Kernel density of the posterior distribution for odds ratio}
\usage{
kernel_plot(result)
}
\arguments{
\item{result}{nof1 result object created using nof1.run}
}
\description{
Kernel density of the posterior distribution for odds ratio
}
