% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nof1.data.R
\name{nof1.data}
\alias{nof1.data}
\title{Make a network object containing data, priors, and a jags model file}
\usage{
nof1.data(Y, Treat, baseline = "baseline", ncat = NULL, response = NULL,
  Time = NULL, knots = NULL, alpha.prior = NULL, beta.prior = NULL,
  gamma.prior = NULL, dc.prior = NULL, c1.prior = NULL,
  rho.prior = NULL, hy.prior = NULL)
}
\arguments{
\item{Y}{Outcome}

\item{Treat}{Treatment indicator vector}

\item{baseline}{baseline Treatment name}

\item{ncat}{Number of categories. Used in ordinal models}

\item{response}{Type of outcome. Can be normal, binomial, poisson or ordinal}
}
\description{
Make a network object containing data, priors, and a jags model file
}
