% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nof1.simulation.R
\name{nof1.ordinal.simulation}
\alias{nof1.ordinal.simulation}
\title{Ordinal simulation}
\usage{
nof1.ordinal.simulation(baseline = "baseline", Base.size = 100,
  Treat.size = 100, alpha = 0, beta_A = -0.1, beta_B = -0.3,
  cut = c(0.5, 1, 1.5, 2), ncat = 5)
}
\description{
Ordinal simulation
}
