% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nof1.simulation.R
\name{nof1.poisson.simulation}
\alias{nof1.poisson.simulation}
\title{Poisson simulation}
\usage{
nof1.poisson.simulation(Base.size = 14, Treat.size = 56, alpha = 1,
  beta_A = -0.1, beta_B = -0.05)
}
\description{
Poisson simulation
}
