% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nof1.run.R
\name{nof1.run}
\alias{nof1.run}
\title{Run nof1 model}
\usage{
nof1.run(nof1, inits = NULL, n.chains = 3, max.run = 1e+05,
  setsize = 10000, n.run = 50000, conv.limit = 1.05,
  extra.pars.save = NULL)
}
\arguments{
\item{nof1}{nof1 object created using nof1.data}
}
\description{
Run nof1 model
}
