% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nof1.data.R
\name{read_input_data}
\alias{read_input_data}
\title{Read json data in as an R object}
\usage{
read_input_data(json.file)
}
\arguments{
\item{json.file}{input data}
}
\description{
Read json data in as an R object
}
