% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nof1.wrap.R
\name{summarize_nof1}
\alias{summarize_nof1}
\title{Summarizes the result from the model into json format}
\usage{
summarize_nof1(nof1, result)
}
\description{
Summarizes the result from the model into json format
}
