% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nof1.summary.R
\name{table_nof1_raw}
\alias{table_nof1_raw}
\title{Summary data table for nof1}
\usage{
table_nof1_raw(nof1)
}
\arguments{
\item{nof1}{nof1 object created using nof1.data}
}
\description{
Summary data table for nof1
}
