% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nof1.summary.R
\name{time_series_plot}
\alias{time_series_plot}
\title{Time series plot for the raw data}
\usage{
time_series_plot(nof1, time = NULL, timestamp = NULL,
  timestamp.format = "\%m/\%d/\%Y \%H:\%M")
}
\arguments{
\item{nof1}{nof1 object created using nof1.data}

\item{time}{can manually specify time variable}

\item{timestamp}{or instead provide timestamp information for the all the outcomes}

<<<<<<< HEAD
\item{timestamp.format}{format of timestamp used. See default format.}
=======
\item{timestamp.format}{format of timestamp used. By default it is and example of it would be 12/24/2015 11:23}
>>>>>>> parent of fe174c7... revision
}
\description{
Draw time series plot
}
