% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nof1.summary.R
\name{time_series_plot2}
\alias{time_series_plot2}
\title{time series plot across different interventions}
\usage{
time_series_plot2(nof1, time = NULL, timestamp = NULL,
  timestamp.format = "\%m/\%d/\%Y \%H:\%M")
}
\arguments{
\item{nof1}{nof1 object created using nof1.data}
}
\description{
time series plot across different interventions
}
