% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nof1.wrap.R
\name{wrap}
\alias{wrap}
\title{Wrapper function that runs the n-of-1 model}
\usage{
wrap(data, metadata)
}
\arguments{
\item{json.file}{input json data}
}
\description{
Wrapper function that runs the n-of-1 model
}
