% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nof1.wrap2.R
\name{wrap2}
\alias{wrap2}
\title{Wrapper function for afib study that runs the n-of-1 model}
\usage{
wrap2(data, metadata)
}
\arguments{
\item{json.file}{input json data}
}
\description{
Wrapper function for afib study that runs the n-of-1 model
}
