% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nof1.summary.R
\name{odds_ratio_plot}
\alias{odds_ratio_plot}
\title{Odds ratio plot for the raw data}
\usage{
odds_ratio_plot(result.list, result.name = NULL, level = 0.95,
  title = NULL)
}
\arguments{
\item{result.list}{list of nof1 results created using nof1.run}

\item{level}{confidence interval level (default is 0.95)}

\item{name}{of the outcomes. If left unspecified, it numbers each result in order of how it is stored in result.list}
}
\description{
Odds ratio plot for the raw data
}
