% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nof1.wrap.R
\name{read_input_data}
\alias{read_input_data}
\title{Read json data in as an R object}
\usage{
read_input_data(data, metadata)
}
\arguments{
\item{json.file}{input data}
}
\description{
Read json data in as an R object
}
