% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nof1.wrap.R
\name{read_input_data}
\alias{read_input_data}
\title{Read json data in as an R object}
\usage{
read_input_data(data, metadata)
}
\arguments{
\item{data}{input data. see sample input.json in the github repo}

\item{metadata}{metadata. see sample input.json in the github repo}
}
\description{
Read json data in as an R object
}
