% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nof1.summary.R
\name{time_series_plot}
\alias{time_series_plot}
\title{Time series plot for the raw data}
\usage{
time_series_plot(nof1, time = NULL, timestamp = NULL,
  timestamp.format = "\%m/\%d/\%Y \%H:\%M")
}
\arguments{
\item{nof1}{nof1 object created using nof1.data}

\item{time}{can manually specify time variable}

\item{timestamp}{or instead provide timestamp information for the all the outcomes}
}
\description{
Draw time series plot
}
