% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nof1.summary.R
\name{time_series_plot}
\alias{time_series_plot}
\title{Time series plot for the raw data}
\usage{
time_series_plot(nof1)
}
\arguments{
\item{nof1}{nof1 object created using nof1.data}
}
\description{
Time series plot for the raw data
}
