% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nof1.wrap.R
\name{wrap}
\alias{wrap}
\title{Make a network object containing data, priors, and a jags model file}
\usage{
wrap(json.file)
}
\arguments{
\item{json.file}{input json data}
}
\description{
Make a network object containing data, priors, and a jags model file
}
